ia/requirements.txt
numpy
requests
llama-cpp-python
llama-index-embeddings-huggingface
chromadb
pysqlite3-binary
flask